Use the new template-based widget rendering
Django 1.11 has removed some undocumented classes in
django.forms.widgets because they are not needed with the new
template-based widget rendering system. Django-colortag used to
depend on those removed classes.

The default HTML templates of the base widgets are used without
changes, but the custom ColortagSelectMultiple widget adds
additional HTML attributes to the elements. The previous version
of the code had some attributes based on the fields of ColorTag
instances, so the new version had to add the ColorTag instances
to the widget.choices attribute that is set by the form field
(custom class ColortagChoiceField). Otherwise, the widget would
not have any access to the model instances themselves.