Fix color tag incl/excl filters after the Django 3.2 upgrade
Django 3.1 added a new instance variable `widgets_names` to
the class MultiWidget. It is a list and if it is empty,
the MultiWidget instance does not render any subwidgets.

In the Jutut feedback platform, feedback and student tags were
missing in the filter feedback form and filtering by tags also
did not work if you manually added the corresponding
GET query parameters to the URL.

The MultiWidget constructor creates default `widgets_names` in
the constructor, but since the tag fields in Jutut are initialized
with empty querysets, the `widgets_names` list was left empty.
We initialize with empty querysets because we know the correct
course instance and its tags only later; they are then filled in.
The solution to the issues is to update the `widgets_names` list
when the querysets and widget choices are filled in afterwards.

It appears that Django's MultiWidget class was designed so that
the widgets would be available already in the constructor.
However, we can just update both the `widgets` and `widgets_names`
instance variables later on.

Fixes apluslms/mooc-jutut#25

https://github.com/apluslms/mooc-jutut/issues/25